GUI/Regression: Disable keyboard listener that stops spamming
In this commit, the keyboard listener responsible for setting the spam value to the final target was disabled. Previously, I wanted the user to press shift to stop the spamming while the spam hadn't finished yet. However, I discovered our method of using keyboard.listener + threading.Thread only works on CLI, but not GUI.py. The exact working mechanism is not known. We temporarily disabled the function to restore the normal functionality of this program. Fixup needed in the future.